+++ OptiScaler/shaders/depth_invert/DI_Dx12.cpp
@@ -60,7 +60,7 @@ static bool CreateComputeShader(ID3D12Device* device, ID3D12RootSignature* rootS
     return true;
 }
 
-bool DI_Dx12::CreateBufferResource(ID3D12Device* InDevice, ID3D12Resource* InSource, uint32_t InWidth,
+bool DI_Dx12::CreateBufferResource(ID3D12Device* InDevice, ID3D12Resource* InSource, uint64_t InWidth,
                                    uint32_t InHeight, D3D12_RESOURCE_STATES InState)
 {
     if (InDevice == nullptr || InSource == nullptr)
