+++ OptiScaler/shaders/depth_invert/DI_Dx12.h
@@ -27,7 +27,7 @@ class DI_Dx12
     D3D12_RESOURCE_STATES _bufferState = D3D12_RESOURCE_STATE_COMMON;
 
   public:
-    bool CreateBufferResource(ID3D12Device* InDevice, ID3D12Resource* InSource, uint32_t InWidth, uint32_t InHeight,
+    bool CreateBufferResource(ID3D12Device* InDevice, ID3D12Resource* InSource, uint64_t InWidth, uint32_t InHeight,
                               D3D12_RESOURCE_STATES InState);
     void SetBufferState(ID3D12GraphicsCommandList* InCommandList, D3D12_RESOURCE_STATES InState);
     bool Dispatch(ID3D12Device* InDevice, ID3D12GraphicsCommandList* InCmdList, ID3D12Resource* InResource,
